add prompt for Yes/No to install packages, programs

+++ .setup/after-setup.sh
@@ -1,2 +1,12 @@
 echo "=========================== copy overwrite ==========================="
-cp -TRv ../ ${ZSH_CUSTOM:-$HOME/}
+while true; do
+    read -p "Do you want copy and overwrite existing config folders from this source to your os? (Y/N)  " yn
+    case $yn in
+    [Yy]*)
+        cp -TRv ../ ${ZSH_CUSTOM:-$HOME/}
+        break
+        ;;
+    [Nn]*) break ;;
+    *) echo "Please answer yes or no." ;;
+    esac
+done
